04-helloagent/requirement.txt
openai-agents>=0.1.0
python-dotenv>=1.0.0
openai>=1.0.0